Fix underflow with vesting computation
One of the errors here is an argument transposition, the other is more
complex, since vesting is only computed on unbonding or claim the value
from the chain may not be updated. This change takes that into account
for at least a common case and adds a ton of asserts to make future
issues easier to find.